requirements.txt
-e .
aws-cdk.aws-codebuild==1.97.0
aws-cdk.core==1.97.0
aws-cdk.aws-lambda==1.97.0
